COMP:Removed old-style cast warning by removing useless cast.

+++ VTK/Infovis/vtkTreeRingPointLayout.cxx
@@ -98,7 +98,7 @@ int vtkTreeRingPointLayout::RequestData( vtkInformation *vtkNotUsed(request),
     }
   }
 
-  double spacing = this->ExteriorRadius / (double)max_level;
+  double spacing = this->ExteriorRadius / max_level;
   
   vtkPoints* points = vtkPoints::New();
   vtkIdType numVerts = outputTree->GetNumberOfVertices();
